feat(🌟Voucher and Points): Events can create an Voucher , but point is not works

+++ src/main/java/com/ilevent/ilevent_backend/pointsHistory/repository/PointHistoryRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PointHistoryRepository extends JpaRepository<PointsHistory, Integer> {
+public interface PointHistoryRepository extends JpaRepository<PointsHistory, Long> {
     List<PointsHistory> findByUserId(Long userId);
 }
